Change of basic IO stream library
Until now the program was using std::iostream as it's mean of communication with user.
However due to difficulties regarding usage of `getch()` and buffer glitches I decided
to switch it back to ncurses entirely.

This switch required to implement new library with is own 'getline' function.
See NI::getline() for more information on how it's work.